+++ app/javascript/lca/components/battlegroups/battlegroupListItem.jsx
@@ -10,7 +10,7 @@ import Typography from 'material-ui/Typography'
 import Launch from 'material-ui-icons/Launch'
 import ModeEdit from 'material-ui-icons/ModeEdit'
 
-import ContentPageCardMenu from '../generic/CharacterCardMenu'
+import CharacterCardMenu from '../generic/CharacterCardMenu'
 import { prettyDrillRating } from '../../utils/calculated'
 
 const styles = theme => ({
@@ -46,7 +46,7 @@ function BattlegroupListItem(props) {
         <Launch className={ classes.icon } />
       </Typography>
 
-      <ContentPageCardMenu characterType="battlegroups" id={ battlegroup.id } />
+      <CharacterCardMenu characterType="battlegroup" id={ battlegroup.id } />
     </div>
 
     <Typography paragraph>