+++ app/javascript/lca/components/generic/CharacterCardMenu/CardMenuRemoveFromChronicle.jsx
@@ -4,52 +4,34 @@ import { connect } from 'react-redux'
 
 import { ListItemIcon, ListItemText } from 'material-ui/List'
 import { MenuItem } from 'material-ui/Menu'
-import Visibility from 'material-ui-icons/Visibility'
-import VisibilityOff from 'material-ui-icons/VisibilityOff'
+import RemoveCircle from 'material-ui-icons/RemoveCircle'
 
-import { updateCharacter, updateQc, updateBattlegroup } from '../../../ducks/actions.js'
+import { removeThingFromChronicle } from '../../../ducks/actions.js'
+import { canIEdit } from '../../../selectors/'
 
-function CardMenuHide(props) {
-  if (!props.show)
+function CardMenuHide({ id, chronId, characterType, canIEdit, removeThing }) {
+  if (!canIEdit || chronId == undefined)
     return <div />
 
-  let action
-  switch(props.characterType) {
-  case 'qcs':
-    action = props.updateQc
-    break
-  case 'battlegroups':
-    action = props.updateBattlegroup
-    break
-  case 'characters':
-  default:
-    action = props.updateCharacter
-  }
-
-  return <MenuItem button>
-    <ListItemText inset primary="Remove Character from Chronicle placeholder" />
-  </MenuItem>
+  return <React.Fragment>
+    <MenuItem button onClick={ () => removeThing(chronId, id, characterType)}>
+      <ListItemIcon><RemoveCircle /></ListItemIcon>
+      <ListItemText inset primary="Remove from Chronicle" />
+    </MenuItem>
+  </React.Fragment>
 }
 CardMenuHide.propTypes = {
+  chronId: PropTypes.number,
   id: PropTypes.number.isRequired,
   characterType: PropTypes.string.isRequired,
-  isHidden: PropTypes.bool,
-  show: PropTypes.bool,
-  updateCharacter: PropTypes.func,
-  updateQc: PropTypes.func,
-  updateBattlegroup: PropTypes.func,
+  canIEdit: PropTypes.bool,
+  removeThing: PropTypes.func,
 }
-function mapStateToProps(state, ownProps) {
-  return {
-    isHidden: state.entities[ownProps.characterType][ownProps.id].hidden,
-    show: state.entities[ownProps.characterType][ownProps.id].chronicle_id != undefined,
-  }
-}
-function mapDispatchToProps(dispatch) {
-  return {
-    updateCharacter:   (id, trait, value) => dispatch(updateCharacter(id, trait, value)),
-    updateQc:          (id, trait, value) => dispatch(updateQc(id, trait, value)),
-    updateBattlegroup: (id, trait, value) => dispatch(updateBattlegroup(id, trait, value)),
-  }
-}
+const  mapStateToProps = (state, ownProps) => ({
+  canIEdit: canIEdit(state, ownProps.id, ownProps.characterType),
+  chronId: state.entities[ownProps.characterType + 's'][ownProps.id].chronicle_id
+})
+const mapDispatchToProps = (dispatch) => ({
+  removeThing: (chronId, id, type) => dispatch(removeThingFromChronicle(chronId, id, type)),
+})
 export default connect(mapStateToProps, mapDispatchToProps)(CardMenuHide)