+++ app/javascript/lca/components/generic/CharacterCardMenu/index.jsx
@@ -1,11 +1,9 @@
 import React from 'react'
 import PropTypes from 'prop-types'
-import { Link } from 'react-router-dom'
 
 import Divider from 'material-ui/Divider'
 import IconButton from 'material-ui/IconButton'
-import { ListItemIcon, ListItemText } from 'material-ui/List'
-import Menu, { MenuItem } from 'material-ui/Menu'
+import Menu from 'material-ui/Menu'
 import MoreVert from 'material-ui-icons/MoreVert'
 
 import CardMenuDelete from './CardMenuDelete.jsx'
@@ -14,7 +12,7 @@ import CardMenuLinks from './CardMenuLinks.jsx'
 import CardMenuPin from './CardMenuPin.jsx'
 import CardMenuRemoveFromChronicle from './CardMenuRemoveFromChronicle.jsx'
 
-class ContentPageCardMenu extends React.Component {
+class CharacterCardMenu extends React.Component {
   constructor(props) {
     super(props)
     this.state = { menuAnchor: null }
@@ -43,8 +41,8 @@ class ContentPageCardMenu extends React.Component {
         onClose={ this.handleClose }
       >
         <CardMenuLinks characterType={ this.props.characterType } id={ this.props.id } />
-        <CardMenuHide characterType={ this.props.characterType } id={ this.props.id } />
         <CardMenuPin characterType={ this.props.characterType } id={ this.props.id } />
+        <CardMenuHide characterType={ this.props.characterType } id={ this.props.id } />
 
         <CardMenuRemoveFromChronicle characterType={ this.props.characterType } id={ this.props.id } />
 
@@ -55,9 +53,9 @@ class ContentPageCardMenu extends React.Component {
     </div>
   }
 }
-ContentPageCardMenu.propTypes = {
+CharacterCardMenu.propTypes = {
   id: PropTypes.number.isRequired,
   characterType: PropTypes.string.isRequired,
 }
 
-export default ContentPageCardMenu
+export default CharacterCardMenu